refactor(projection_mpo): inline DFA utilities and update docs/tests
- move DFA and tensor helper code into projection_mpo.jl
- reorganize related test file
